move options into core
